Added "Include" lists! Tons of cleanup needed
"I want all the streamers for League of Legends and Overwatch, but also give me Kitboga" is the use case.

Really dirty implementation, but it works.

Note that right now, it shows all of your Included streamers first, followed by the rest of the list, so it isn't purely by viewers