0.2.5 - Mon, Aug 27, 2018 - Minor tweaks
Overview
There were a handful of small things that I really wanted to tackle here.

Added
Saving Twitch Bearer token in file, and reading from it.
"Generated At" timestamp on the client. It's interesting to know when the data was last updated.
A common function for dev-mode logging.
A singular function for Twitch API calls. That way we can appropriately log in just one place.
Log every Twitch call made.
Timestamp for server restart.

Changed
/streams/game -> /streams/list. This more accurately reflects what the point of that endpoint is.
/streams/list can handle being passed game IDs and streamer IDs/names decently.
"Bottom buttons" on client. Now are a set of links with relevant things.

Removed
/streams/details endpoint. This was a redundant one, created due to a restriction with the Twitch API. I had found a workaround, so it no longer is necessary