0.3.0 - Tues, Aug 28, 2018 - The client technically works now! (#64)
* Initial 'development' commit

Changing travis.yml, making sure Travis picks up on this branch.

* Adding some more links at the bottom

Converted to pure links (from bootstrap buttons) for simplicity.

* Added basic timestamp for data

* Updating /streams/games to call an external function

This function will be needed for the upcoming improvement.

* Saving twitch bearer token in file

Previously, every time the server got restarted, we lost the memory of our bearer token, and re-generated one.
That seems like a bad idea.
Now we try to read from file, and save it to file as needed.

* Changing /streams/games to /streams/list

the /list name kind of lacks meaning, it's way too specific for what we're actually doing here.

* Cleaned up getting Streams

Twitch API allows you to either specify game or user (if you do both, it acts like an inner join, which we don't want)
So we have to add in a case to handle the user passing both -- by just making two separate Twitch calls and combining them.

* Small logging updates

Just wanted a log statement whenever a call to Twitch gets made, and also a timestamp of when server got restarted.

* Logging cleanup; made a singular axios request function

- Moved logging over to a utils function that can differentiate dev vs. prod logging
- Made a singular function for calling to Twitch (used by post and get requests), that way there can be a singular logging location for every time a request is made.

* Dev logging cleanup

For some reason, `process.env.NODE_ENV || 'dev''` wasn't getting the right result, negating the whole point of what was happening.

* Updating version.

* Fixed an issue with running options through getValidQueryOptions multiple times

There are cases where we pass qs options down multiple levels, re-checking getValidQueryOptions.
A problem arose with boolean elements where we were always treating them like a string that needed to get converted. But that caused an error if they already were booleans.

* Added common functions for merging arrays, and avoiding duplicates

Since this is something that has to get done in multiple places, it's been split into functions.
- removeArrayDuplicates removes them from an individual array
- combineArraysWithoutDuplicates merges and avoids duplicates.

* Updating exports for game & stream modules

These will be needed for the combo search.

* Trying to rely on variables in Test cases more

Adding more games to the commonGames array and utilizing that.

* Testing against duplicates in game results

Making sure we do not get duplicate games in case of a game being part of:
- the top games
- list of games passed by name
- list of games passed by ID

* Trying out making use of 'before' flag to combine tests

First attempt with the /combo endpoint. Will try merging more shortly

* Actually fixed an issue with boolean query options

Turns out we're not trying to "return" here.

* Combining test service calls where possible

This should reduce the spam to Twitch, and reduce the time to run tests.
Plenty of tests require their own separate call, but plenty more were able to be combined to use a single call.

* Removing reference to /api/combo temporarily

This wasn't ready to go out yet, and is breaking Travis.

* Adding in api/combo

This is intended to be the main api call. It will pull both streams and games from Twitch.

* Back on client-side, calling the combo api call

Using the generated time to make sure this call is working for now, before moving on

* Displaying streams!

Getting back to where we used to be, yea boi

* Updating version