Add a way to get missing games details in the Combo call
Doing the combo search gets both Game and Stream details. And the main flow gets Games first, and Streams after. But if you specify certain Streams to get, they might not be playing something in your existing Games list, causing you to not have any data on that game.

This will check for that case, and request game information on those necessary games before returning.