0.4.0 - Mon, Sept 3, 2018 - Searching, and hooking up React events (#68)
* Initial 'development' commit

Changing travis.yml, making sure Travis picks up on this branch.

* Adding some more links at the bottom

Converted to pure links (from bootstrap buttons) for simplicity.

* Added basic timestamp for data

* Updating /streams/games to call an external function

This function will be needed for the upcoming improvement.

* Saving twitch bearer token in file

Previously, every time the server got restarted, we lost the memory of our bearer token, and re-generated one.
That seems like a bad idea.
Now we try to read from file, and save it to file as needed.

* Changing /streams/games to /streams/list

the /list name kind of lacks meaning, it's way too specific for what we're actually doing here.

* Cleaned up getting Streams

Twitch API allows you to either specify game or user (if you do both, it acts like an inner join, which we don't want)
So we have to add in a case to handle the user passing both -- by just making two separate Twitch calls and combining them.

* Small logging updates

Just wanted a log statement whenever a call to Twitch gets made, and also a timestamp of when server got restarted.

* Logging cleanup; made a singular axios request function

- Moved logging over to a utils function that can differentiate dev vs. prod logging
- Made a singular function for calling to Twitch (used by post and get requests), that way there can be a singular logging location for every time a request is made.

* Dev logging cleanup

For some reason, `process.env.NODE_ENV || 'dev''` wasn't getting the right result, negating the whole point of what was happening.

* Updating version.

* Fixed an issue with running options through getValidQueryOptions multiple times

There are cases where we pass qs options down multiple levels, re-checking getValidQueryOptions.
A problem arose with boolean elements where we were always treating them like a string that needed to get converted. But that caused an error if they already were booleans.

* Added common functions for merging arrays, and avoiding duplicates

Since this is something that has to get done in multiple places, it's been split into functions.
- removeArrayDuplicates removes them from an individual array
- combineArraysWithoutDuplicates merges and avoids duplicates.

* Updating exports for game & stream modules

These will be needed for the combo search.

* Trying to rely on variables in Test cases more

Adding more games to the commonGames array and utilizing that.

* Testing against duplicates in game results

Making sure we do not get duplicate games in case of a game being part of:
- the top games
- list of games passed by name
- list of games passed by ID

* Trying out making use of 'before' flag to combine tests

First attempt with the /combo endpoint. Will try merging more shortly

* Actually fixed an issue with boolean query options

Turns out we're not trying to "return" here.

* Combining test service calls where possible

This should reduce the spam to Twitch, and reduce the time to run tests.
Plenty of tests require their own separate call, but plenty more were able to be combined to use a single call.

* Removing reference to /api/combo temporarily

This wasn't ready to go out yet, and is breaking Travis.

* Adding in api/combo

This is intended to be the main api call. It will pull both streams and games from Twitch.

* Back on client-side, calling the combo api call

Using the generated time to make sure this call is working for now, before moving on

* Displaying streams!

Getting back to where we used to be, yea boi

* Updating version

* Add links in readme to the app and to changelog

* Add ESLint once again, with some less-strict rules

* Add eslint/prettier rules for more personal usage

* Add a version number to the client

This helps figure out what functionality should/n't be present.

* Start rework of QueryOption parsing

Taking a much more modular approach to this instead of having the behemoth function.

* Add better type-checking for queryoptions

* Add testing for array simplification

* Add testing for falling back to default values

* Finish building the replacement function for QueryOptions

This function is a much cleaner way of parsing what the user has put in, and returning only the relevant and valid members.

* Add conversion to Twitch-specific parameters

* Remove old QueryOptions code

* Add a way to map from Twitch endpoints to our own

* Switch Games over to the new query parsing methodology

* Complete conversion to new QueryOptions system

- Completely rewrote the process for interpreting querystring values
- Decoupled the incoming values from outgoing Twitch ones, particularly to reduce ambiguity (e.g. "id" could be stream or game ID)
- Added a function to convert from internal options to Twitch options, done solely at the time of making a Twitch request.
- Updated all usage of query params.
- Added and updated all tests accordingly
- Moved qs parsing to the most generic parts of calls as possible

* Add a way to get missing games details in the Combo call

Doing the combo search gets both Game and Stream details. And the main flow gets Games first, and Streams after. But if you specify certain Streams to get, they might not be playing something in your existing Games list, causing you to not have any data on that game.

This will check for that case, and request game information on those necessary games before returning.

* Add click handler in OptionsPane

Right now this just cares about the base case of the two checkboxes, but this should be extensible to all others.

* Upgrade Babel to 7

* Upgrade to Babel 7

Missed this file.

* Implement React-Select for Games/Include/Exclude

Replaced the 4 big elements (List of games, Additional games, Include users, Exclude users) with 3 React-Selects (https://react-select.com/). These are tag-based Selects, that allow for default values to be declared, and new ones to be added.
A lot of background things got shifted around to make this happen.

* Implement handler for updating Include/Exclude lists

This common handler in App will take care of all 3 of those Selects, appropriately updating the state.

* Implement search functionality

Hooked up the Search button to actually pull from the State, and update the directory

* Add labels and spacing in OptionsPane

Minor visual updates

* Minor cleanup in Directory

* Minor cleanup in App

* Update version number