Animate the generated time on update
This gives just a brief animation (flashing background color) to the generated time number whenever we get new data from the server.

The loading indicator only shows up if the request is taking a while; this complements that functionality by always showing, but being a bit more out of the way.