Fix animation triggers for updated times
Previously I thought the only time a re-render would get triggered is if the time was new. So I used that as the logic to animate.  That was a wrong assumption.
Now we compare the existing vs. new times, and set an animation flag. If the flag is true, we animate.
Upon ending the animation, we set the flag to false.