Allow broader set of language results
#93
Previously, searching for English streams only returned those with "en" as the language. Streamers with "en-gb" as their language got filtered out. Now things are a little bit more flexible.

This is kind of a band-aid until Twitch adds Tags to their API.
This is also an edge case for the most part. Very few people stream with en-gb. The biggest scenario is probably pt-br instead of pt.