Fully hide the loading indicator after animating
Fixes #95 . Once the opacity animation is finished, set visibility to hidden.
Without that, the loading indicator was still taking up room and taking up clicks intended for a stream cell.

Also splitting the shorthand "transition" property out, for ease of understanding, since I could never really remember which was duration and which was delay.